establish register system and relationship between user and tasks

+++ app/routes/tasks.py
@@ -9,18 +9,18 @@ tasks_bp = Blueprint('tasks',__name__)
 def view_tasks():
   if 'user' not in session:
     return redirect(url_for('auth.login'))
-  
-  tasks = Task.query.all()
+  username = session['user']
+  tasks = Task.query.filter_by(user_id=username).all()
   return render_template('tasks.html',tasks=tasks)
 
 @tasks_bp.route('/add',methods=['POST'])
 def add_task():
   if 'user' not in session:
     return redirect(url_for('auth.login'))
-  
+  username = session['user']
   title = request.form.get('title')
   if title:
-    new_task = Task(title=title,status='Pending')
+    new_task = Task(title=title,status='Pending',user_id=username)
     db.session.add(new_task)
     db.session.commit()
     flash('Task Added Successfully','success')
@@ -41,7 +41,8 @@ def toggle_status(task_id):
 
 @tasks_bp.route('/clear',methods=['POST'])
 def clear_tasks():
-  Task.query.delete()
+  username = session['user']
+  Task.query.filter_by(user_id=username).delete()
   db.session.commit()
   flash('ALL TASKS CLEARED','info')
   return redirect(url_for('tasks.view_tasks'))
@@ -49,12 +50,11 @@ def clear_tasks():
 @tasks_bp.route('/clear_task/<int:task_id>', methods=['POST'])
 def delete_task(task_id):
   task = Task.query.get(task_id)
-  if task:
+  if task and task.user_id == session.get('user'):
     db.session.delete(task)
     db.session.commit()
     flash("Task Deleted Successfully",'info')
     
     flash('Task does not exist')
-    flash('task does not exists')
   return redirect(url_for('tasks.view_tasks'))
